feat(sync): integrate spec kit plugin for synchronization with issue tracking platforms
- add configuration file for sync plugin
- implement sync command with push, pull, and status subcommands
- support for GitHub, Jira, and Asana platforms
- enhance CLI with auto-sync and config management options
