+++ example/.specify/config.yml
@@ -1,0 +1,12 @@
+version: "1.0"
+
+plugins:
+  sync:
+    platform: github
+    autoSync: true
+    conflictStrategy: manual
+    github:
+      owner: ${GITHUB_OWNER}
+      repo: ${GITHUB_REPO}
+      auth: cli
+      # token: ${GITHUB_TOKEN}  # Optional: for token-based auth
