+++ package.json
@@ -16,7 +16,7 @@
     "test": "turbo run test",
     "test:coverage": "turbo run test:coverage"
   },
-  "packageManager": "bun@1.2.20",
+  "packageManager": "bun@1.2.22",
   "workspaces": [
     "apps/*",
     "packages/*",
