+++ packages/cli/src/commands/sync.ts
@@ -1,338 +1,18 @@
 /**
  * Sync command - Synchronize specs with issue tracking platforms
+ * Now integrated with @spec-kit/plugin-sync
  */
 
 import type { Command } from 'commander'
-import process from 'node:process'
-import pc from 'picocolors'
-import { consoleUtils } from '../ui/Console.js'
-import { InteractiveSelect } from '../ui/InteractiveSelect.js'
-import { StepTracker } from '../ui/StepTracker.js'
+import { createSyncCommand } from '@spec-kit/plugin-sync'
 
-// Import sync plugin modules (these will be available after dependency is added)
-interface SyncOptions {
-  force?: boolean
-  dryRun?: boolean
-  verbose?: boolean
-  platform?: string
-  conflictStrategy?: 'manual' | 'theirs' | 'ours' | 'interactive'
-}
+export function registerSyncCommands(program: Command) {
+  // Use the plugin's sync command factory
+  const syncCommand = createSyncCommand({
+    name: 'sync',
+    description: 'Synchronize specs with issue tracking platforms',
+  })
 
-// Platform options
-const PLATFORMS = ['github', 'jira', 'asana'] as const
-const CONFLICT_STRATEGIES = ['manual', 'theirs', 'ours', 'interactive'] as const
-
-export async function registerSyncCommands(program: Command): Promise<void> {
-  // Main sync command with subcommands
-  const syncCmd = program
-    .command('sync')
-    .description('Synchronize specs with issue tracking platforms')
-    .option('--platform <platform>', `Platform to sync with (${PLATFORMS.join(', ')})`, 'github')
-    .option('--config <path>', 'Path to configuration file')
-    .option('--verbose', 'Enable verbose logging', false)
-    .option('--dry-run', 'Preview changes without applying them', false)
-    .addHelpText('after', `
-Examples:
-  $ specify sync status
-  $ specify sync push --all
-  $ specify sync pull 123
-  $ specify sync --platform github push specs/001-feature
-  $ specify sync --platform jira status
-  $ specify sync config --show
-
-Available platforms:
-  ${PLATFORMS.map(p => `  ${p.padEnd(8)} - ${getPlatformDescription(p)}`).join('\n')}
-
-Configuration:
-  Create .specify/sync.config.json with platform credentials
-  Run 'specify sync config --show' to see current configuration
-`)
-
-  // Push subcommand
-  syncCmd
-    .command('push [spec-path]')
-    .description('Push specs to remote platform')
-    .option('-a, --all', 'Sync all specs')
-    .option('-f, --force', 'Force sync even if no changes detected')
-    .option('--conflict-strategy <strategy>', `Conflict resolution strategy (${CONFLICT_STRATEGIES.join(', ')})`, 'manual')
-    .action(async (specPath, options, command) => {
-      const globalOptions = command.parent?.opts() || {}
-      await handleSyncPush(specPath, options, globalOptions)
-    })
-
-  // Pull subcommand
-  syncCmd
-    .command('pull [issue-id]')
-    .description('Pull issues from remote platform to specs')
-    .option('-a, --all', 'Pull all issues with spec label')
-    .option('-f, --force', 'Force overwrite existing specs')
-    .action(async (issueId, options, command) => {
-      const globalOptions = command.parent?.opts() || {}
-      await handleSyncPull(issueId, options, globalOptions)
-    })
-
-  // Status subcommand
-  syncCmd
-    .command('status')
-    .description('Check sync status of all specs')
-    .action(async (command) => {
-      const globalOptions = command.parent?.opts() || {}
-      await handleSyncStatus(globalOptions)
-    })
-
-  // Config subcommand
-  syncCmd
-    .command('config')
-    .description('Manage sync configuration')
-    .option('--show', 'Show current configuration')
-    .option('--platform <platform>', 'Set default platform')
-    .action(async (options, command) => {
-      const globalOptions = command.parent?.opts() || {}
-      await handleSyncConfig(options, globalOptions)
-    })
-}
-
-async function handleSyncPush(specPath: string | undefined, options: any, globalOptions: any): Promise<void> {
-  try {
-    consoleUtils.log(pc.cyan('🔄 Syncing specs to platform...'))
-    consoleUtils.log()
-
-    const syncOptions: SyncOptions = {
-      force: options.force,
-      dryRun: globalOptions.dryRun,
-      verbose: globalOptions.verbose,
-      platform: globalOptions.platform,
-      conflictStrategy: options.conflictStrategy,
-    }
-
-    // Validate platform
-    if (!PLATFORMS.includes(syncOptions.platform as any)) {
-      const selectedPlatform = await selectPlatform(syncOptions.platform)
-      syncOptions.platform = selectedPlatform
-    }
-
-    // Create step tracker for progress
-    const steps = [
-      'Loading configuration',
-      'Authenticating with platform',
-      'Scanning specifications',
-      options.all ? 'Syncing all specs' : `Syncing ${specPath || 'current spec'}`,
-      'Updating metadata',
-    ]
-
-    const tracker = new StepTracker('Syncing Specifications')
-
-    // Add steps to tracker
-    steps.forEach(step => tracker.add(step, step))
-    tracker.start('Loading configuration')
-    consoleUtils.log(tracker.render())
-
-    // Simulate sync operations (replace with actual plugin calls)
-    await simulateSync(tracker, syncOptions, specPath, options.all)
-  }
-  catch (error) {
-    consoleUtils.error('Sync push failed', error instanceof Error ? error : new Error(String(error)))
-    process.exit(1)
-  }
-}
-
-async function handleSyncPull(issueId: string | undefined, options: any, globalOptions: any): Promise<void> {
-  try {
-    consoleUtils.log(pc.cyan('📥 Pulling issues from platform...'))
-    consoleUtils.log()
-
-    const syncOptions: SyncOptions = {
-      force: options.force,
-      dryRun: globalOptions.dryRun,
-      verbose: globalOptions.verbose,
-      platform: globalOptions.platform,
-    }
-
-    // Validate platform
-    if (!PLATFORMS.includes(syncOptions.platform as any)) {
-      const selectedPlatform = await selectPlatform(syncOptions.platform)
-      syncOptions.platform = selectedPlatform
-    }
-
-    if (globalOptions.dryRun) {
-      consoleUtils.log(pc.dim('[DRY RUN] Would pull issues from platform'))
-      return
-    }
-
-    consoleUtils.log(pc.green(`✓ Would pull issue ${issueId || 'all issues'} from ${syncOptions.platform}`))
-    consoleUtils.log(pc.dim('Pull functionality will be available after plugin integration'))
-  }
-  catch (error) {
-    consoleUtils.error('Sync pull failed', error instanceof Error ? error : new Error(String(error)))
-    process.exit(1)
-  }
-}
-
-async function handleSyncStatus(globalOptions: any): Promise<void> {
-  try {
-    consoleUtils.log(pc.cyan('📊 Checking sync status...'))
-    consoleUtils.log()
-
-    const platform = globalOptions.platform || 'github'
-
-    // Validate platform
-    if (!PLATFORMS.includes(platform as any)) {
-      const selectedPlatform = await selectPlatform(platform)
-      globalOptions.platform = selectedPlatform
-    }
-
-    consoleUtils.log(pc.blue(`Spec Sync Status (${globalOptions.platform}):`))
-    consoleUtils.log(pc.gray('─'.repeat(60)))
-    consoleUtils.log()
-
-    // Simulate status display (replace with actual plugin calls)
-    const mockSpecs = [
-      { name: '001-user-authentication', issue: 123, status: 'synced', hasChanges: false },
-      { name: '002-payment-processing', issue: 124, status: 'draft', hasChanges: true },
-      { name: '003-data-analytics', issue: null, status: 'draft', hasChanges: true },
-    ]
-
-    for (const spec of mockSpecs) {
-      const statusColor
-        = spec.status === 'synced'
-          ? pc.green
-          : spec.status === 'conflict'
-            ? pc.red
-            : spec.status === 'draft'
-              ? pc.yellow
-              : pc.gray
-
-      const changeIndicator = spec.hasChanges ? '●' : '○'
-
-      consoleUtils.log(
-        `${changeIndicator} ${pc.cyan(spec.name.padEnd(25))} `
-        + `Issue: ${spec.issue ? pc.white(`#${spec.issue}`) : pc.gray('none')} `
-        + `Status: ${statusColor(spec.status)}`,
-      )
-    }
-
-    consoleUtils.log()
-    consoleUtils.log(pc.dim('Run "specify sync push --all" to sync pending changes'))
-  }
-  catch (error) {
-    consoleUtils.error('Sync status failed', error instanceof Error ? error : new Error(String(error)))
-    process.exit(1)
-  }
-}
-
-async function handleSyncConfig(options: any, globalOptions: any): Promise<void> {
-  try {
-    if (options.show) {
-      consoleUtils.log(pc.cyan('⚙️ Current sync configuration:'))
-      consoleUtils.log()
-
-      const mockConfig = {
-        platform: globalOptions.platform || 'github',
-        autoSync: true,
-        conflictStrategy: 'manual',
-        github: {
-          owner: 'your-org',
-          repo: 'your-repo',
-          auth: 'cli',
-        },
-      }
-
-      consoleUtils.log(JSON.stringify(mockConfig, null, 2))
-      consoleUtils.log()
-      consoleUtils.log(pc.dim('Configuration file: .specify/sync.config.json'))
-    }
-    else if (options.platform) {
-      consoleUtils.log(pc.yellow('Configuration management not yet implemented'))
-      consoleUtils.log(pc.dim(`Would set default platform to: ${options.platform}`))
-    }
-    else {
-      consoleUtils.log(pc.cyan('Sync Configuration Management'))
-      consoleUtils.log()
-      consoleUtils.log('Options:')
-      consoleUtils.log('  --show              Show current configuration')
-      consoleUtils.log('  --platform <name>   Set default platform')
-      consoleUtils.log()
-      consoleUtils.log('Example:')
-      consoleUtils.log('  specify sync config --show')
-      consoleUtils.log('  specify sync config --platform github')
-    }
-  }
-  catch (error) {
-    consoleUtils.error('Config command failed', error instanceof Error ? error : new Error(String(error)))
-    process.exit(1)
-  }
-}
-
-async function selectPlatform(currentPlatform?: string): Promise<string> {
-  consoleUtils.log(pc.yellow(`Platform "${currentPlatform}" not recognized.`))
-  consoleUtils.log('Please select a platform:')
-  consoleUtils.log()
-
-  const platformOptions = PLATFORMS.map(platform => ({
-    name: `${platform} - ${getPlatformDescription(platform)}`,
-    value: platform,
-  }))
-
-  const platformMap = platformOptions.reduce((acc, option) => {
-    acc[option.value] = option.name
-    return acc
-  }, {} as Record<string, string>)
-
-  return await InteractiveSelect.select(
-    platformMap,
-    'Select platform:',
-  )
-}
-
-function getPlatformDescription(platform: string): string {
-  switch (platform) {
-    case 'github':
-      return 'GitHub Issues & Projects'
-    case 'jira':
-      return 'Atlassian Jira (coming soon)'
-    case 'asana':
-      return 'Asana Tasks & Projects (coming soon)'
-    default:
-      return 'Unknown platform'
-  }
-}
-
-async function simulateSync(tracker: StepTracker, options: SyncOptions, specPath?: string, syncAll?: boolean): Promise<void> {
-  // Step 1: Loading configuration
-  await new Promise(resolve => setTimeout(resolve, 500))
-  tracker.complete('Loading configuration')
-  tracker.start('Authenticating with platform')
-
-  // Step 2: Authentication
-  await new Promise(resolve => setTimeout(resolve, 800))
-  tracker.complete('Authenticating with platform')
-  tracker.start('Scanning specifications')
-
-  // Step 3: Scanning specs
-  await new Promise(resolve => setTimeout(resolve, 600))
-  tracker.complete('Scanning specifications')
-
-  const syncStepName = syncAll ? 'Syncing all specs' : `Syncing ${specPath || 'current spec'}`
-  tracker.start(syncStepName)
-
-  // Step 4: Syncing
-  await new Promise(resolve => setTimeout(resolve, 1200))
-  tracker.complete(syncStepName)
-  tracker.start('Updating metadata')
-
-  // Step 5: Updating metadata
-  await new Promise(resolve => setTimeout(resolve, 400))
-  tracker.complete('Updating metadata')
-
-  consoleUtils.log(tracker.render())
-  consoleUtils.log()
-
-  if (options.dryRun) {
-    consoleUtils.log(pc.yellow('🔍 DRY RUN - No changes were made'))
-    consoleUtils.log(pc.dim(`Would sync to platform: ${options.platform}`))
-  }
-  else {
-    consoleUtils.log(pc.green('✅ Sync completed successfully'))
-    consoleUtils.log(pc.dim('Note: Full sync functionality will be available after plugin integration'))
-  }
+  // Add the sync command to the main program
+  program.addCommand(syncCommand)
 }
