+++ plugins/sync/eslint.config.mjs
@@ -3,5 +3,5 @@ import antfu from '@antfu/eslint-config'
 export default antfu({
   formatters: true,
 }, {
-
+    'no-console': 'off',
 })
