+++ plugins/sync/package.json
@@ -5,23 +5,26 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "bin": {
-    "specify-sync": "./dist/cli.js"
+    "specify-sync": "./dist/cli"
   },
   "scripts": {
-    "build": "bun build ./src/cli.ts --compile --outfile ./dist/cli ",
+    "build": "bun run clean && bun run build:lib && bun run build:cli",
+    "build:lib": "bun build ./src/index.ts --target=bun --outdir=./dist --splitting && tsc -p tsconfig.build.json",
+    "build:cli": "bun build ./src/cli.ts --compile --outfile ./dist/cli",
     "dev": "bun build ./src/cli.ts --compile --outfile ./dist/cli --watch",
+    "dev:lib": "bun build ./src/index.ts --target=bun --outdir=./dist --splitting --watch",
     "clean": "rm -rf dist",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "typecheck": "tsc --noEmit",
     "test": "bun test",
     "test:coverage": "bun test --coverage",
-    "specify-sync": "bun run dist/cli.js"
+    "specify-sync": "bun run dist/cli"
   },
   "dependencies": {
     "@spec-kit/core": "workspace:*",
     "chalk": "^5.3.0",
-    "commander": "^12.1.0",
+    "commander": "^14.0.1",
     "gray-matter": "^4.0.3",
     "zod": "^4.1.8"
   },
