+++ plugins/sync/src/types/index.ts
@@ -2,6 +2,7 @@ import type { SpecFileFrontmatter } from '../schemas'
 import { z } from 'zod'
 
 export type { SpecFileFrontmatter } from '../schemas/spec.ts'
+export type { SyncAdapter, RemoteRef, AdapterCapabilities } from '../adapters/base.adapter.js'
 
 export interface SpecFile {
   path: string
