+++ plugins/sync/tsconfig.json
@@ -22,6 +22,7 @@
     "noUnusedLocals": false,
     "noUnusedParameters": false,
     "noEmit": true,
+    "declaration": true,
     "verbatimModuleSyntax": true,
     "skipLibCheck": true
   },
