--- plugins/sync/tsconfig.types.json
@@ -1,13 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "noEmit": false,
-    "outDir": "./dist"
-  },
-  "include": [
-    "src/**/*.ts"
-  ]
-}
